Added Login Input Field and Input field to seprate files.
