+++ src/app/component/EditProfile.tsx
@@ -5,6 +5,7 @@ const EditProfile = () => {
     <div className="flex flex-col">
       <h1 className="text-2xl ml-1">Edit Profile</h1> {/* divider */}
       <div className="bg-dusk h-0.5 w-4xl mb-2 mt-2"></div>
+      {/* Warning message from WarningMessage.tsx */}
       <WarningMessage
         containerClassName="w-xl"
         textClassName=" "
