+++ src/app/component/RegistrationForm.tsx
@@ -2,18 +2,7 @@
 
 import WarningMessage from "@/app/component/WarningMessage";
 import { useState } from "react";
-
-const InputField = ({ label, type, name, placeholder, className = "" }) => (
-  <div className="flex flex-col w-full">
-    <label className="text-sm text-gray-300 mb-1">{label}</label>
-    <input
-      type={type}
-      name={name}
-      placeholder={placeholder}
-      className={`px-3 py-2 bg-black/30 border outline-2 rounded-md text-white focus:outline-none focus:ring-2 focus:ring-teal ${className}`}
-    />
-  </div>
-);
+import InputField from "./InputField";
 
 export default function Form() {
   const [selectedCourse, setSelectedCourse] = useState("");
@@ -147,6 +136,7 @@ export default function Form() {
           />
         </div>
 
+        {/* Warning message from WarningMessage.tsx */}
         <WarningMessage
           containerClassName="col-span-2"
           textClassName=" "
